requirements.txt
asgiref==3.7.2
Babel==2.14.0
cffi==1.16.0
cryptography==41.0.7
Django==4.2.4
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-money==3.4.1
django-mptt==0.15.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
phonenumbers==8.13.29
Pillow==10.1.0
py-moneyed==3.0
pycparser==2.21
pytz==2023.3.post1
setuptools==69.0.3
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
